requirements.txt
azure-identity
msgraph-sdk
azure-functions
azure-mgmt-resourcegraph
